완성

+++ teamVoca/routes/commu.js
@@ -33,13 +33,10 @@ router.get("/vocas", async (req, res) => {
       },
     ],
   });
+
   // return res.json(row);
   // return res.json({ rows, row, userID });
-  return res.render("commu/community", {
-    VOCAS: rows,
-    LIKE: row,
-    user: userID,
-  });
+  return res.render("commu/community", { VOCAS: rows, LIKE: row, user: userID });
 });
 
 router.get("/:v_seq/like", async (req, res) => {
@@ -53,9 +50,8 @@ router.get("/:v_seq/like", async (req, res) => {
 
   let like = voca.v_rec;
 
-  const rec = await LIKE.findAll({
-    where: { like_user: userID, like_vseq: v_seq },
-  });
+
+  const rec = await LIKE.findAll({ where: { like_user: userID, like_vseq: v_seq } });
   if (rec.length === 0) {
     req.body.like_user = userID;
     req.body.like_vseq = v_seq;
@@ -63,9 +59,8 @@ router.get("/:v_seq/like", async (req, res) => {
     like = v_rec + 1;
     await VOCAS.update({ v_rec: like }, { where: { v_seq: v_seq } });
   } else {
-    await LIKE.destroy({
-      where: { like_user: userID, like_vseq: v_seq },
-    });
+    await LIKE.destroy({ where: { like_user: userID, like_vseq: v_seq } });
+
 
     like = v_rec - 1;
     await VOCAS.update({ v_rec: like }, { where: { v_seq: v_seq } });
@@ -87,12 +82,10 @@ router.get("/:v_seq/like", async (req, res) => {
     ],
   });
   // return res.json(row);
-  return res.render("commu/community", {
-    VOCAS: rows,
-    LIKE: row,
-    user: userID,
-  });
-  // return res.redirect("/commu/vocas");
+
+  // return res.render("commu/community", { VOCAS: rows, LIKE: row, user: userID });
+  return res.redirect("/commu/vocas");
+
 });
 
 router.get("/:v_seq/detail", async (req, res) => {
@@ -117,12 +110,9 @@ router.get("/:v_seq/detail", async (req, res) => {
   });
   const user = req.session.user ? req.session.user.m_id : undefined;
 
-  return res.render("commu/detail", {
-    rows,
-    voca,
-    COMMENT: comment,
-    user,
-  });
+
+  return res.render("commu/detail", { rows, voca, COMMENT: comment, user });
+
 });
 
 router.post("/:v_seq/detail", async (req, res) => {
